mcp_server/requirements.txt
mcp>=0.4.0
fastmcp>=0.4.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0